requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
blockbuster==1.5.25
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
cryptography==44.0.3
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
exa-py==1.15.2
executing==2.2.0
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.7.0
gcloud==0.18.3
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.110.0
google-cloud-bigquery==3.36.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.31.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.75
langchain-exa==0.3.0
langchain-google-genai==2.1.10
langchain-google-vertexai==2.0.28
langchain-tavily==0.2.11
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-api==0.3.1
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.8
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.8.1
langgraph-sdk==0.2.3
langsmith==0.4.16
matplotlib-inline==0.1.7
multidict==6.6.4
nest-asyncio==1.6.0
nltk==3.9.1
numexpr==2.11.0
numpy==2.3.2
oauth2client==4.1.3
openai==1.101.0
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.47.2
structlog==25.4.0
tenacity==9.1.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
